Fix bind to create new window in $PWD
The previous bind didn't work if, say, you have two panes in the current
window. If the first window has neovim opened, and you enter the keybind
in the second window, then no new window is created. Instead, the neovim
pane is updated with parts of the keybind text.

The reason is that `-t.-` (roughly speaking) refers to "the other pane".